* bebanjo: (32 commits)
  can restart build with cancelled jobs
  fixed job_updater: set to deploy state only if regular jobs successfuly completed
  add prefix to production env file
  remove secret token from repo
  update vx-instrumentation gem
  Using UUID as pk in invites
  ui fixes
  add bs-callout css class
  Company usage statistics
  fixed current user menu
  Can destroy users
  Add current company scope to SSE
  DRY Gemfile and database.yml
  new deploy
  new deploy
  update vx-consumer gem
  update vx-consumer gem
  remove force ssl option
  add force_ssl to application config
  add ajax loaded to /ui/users
  ...

Conflicts:
	Gemfile
	Gemfile.lock
	app/models/build.rb
	app/models/job.rb
	app/models/job_updater.rb
	db/schema.rb
	spec/models/build_spec.rb

+++ app/controllers/users/github_controller.rb
@@ -34,14 +34,13 @@ class Users::GithubController < ApplicationController
     end
 
     def invite
-      company = Company.find_by! name: company_name
-      invite  = company.invites.find_by! token: token, email: email
+      invite  = Invite.find_by! token: token, id: invite_id
+      company = invite.company
+      email   = invite.email
       github  = UserSession::Github.new request.env["omniauth.auth"]
       user    = github.create(email, trust_email: true)
 
-      if user.valid?
-        # TODO: check add_to_company result
-        user.add_to_company company
+      if user.valid? and user.add_to_company(company, invite.role)
         invite.destroy
         session[:user_id] = user.id
         redirect_to_saved_location_or_root
@@ -54,16 +53,12 @@ class Users::GithubController < ApplicationController
       request.env['omniauth.params']
     end
 
-    def email
-      o_params["email"]
-    end
-
     def token
-      o_params["token"]
+      o_params["t"]
     end
 
-    def company_name
-      o_params["company"]
+    def invite_id
+      o_params["i"]
     end
 
 end
